niche/generate-docs.bashlib: avoid extra doc/man deps if possible
Instead of setting USE="doc man" for the initial emerge unconditionally,
only do it if the "prebuilt doc infrastructure" isn't present in the ebuild.

If it is present, we can just say the docs aren't prebuilt yet
(which they won't be for a new version anyway) and install
those docs as a direct dep, rather than needing to set USE flags
which if set on the command line will apply transitively to deps-of-deps.

Signed-off-by: Sam James <sam@gentoo.org>