maint/bump-go: add recursive go.mod support
dev-db/etcd has several subdirs with go.mod files. We want to recursively
grab all of those. It's easiest to just use a shared go-mod dir (hence now using
${WORKDIR}/go-mod and not ${PWD}/go-mod) for that.

This is only supported for "download" ("dep tarballs") rather than vendor tarballs
(which is something I rarely test as it's not super common that it even works).

Signed-off-by: Sam James <sam@gentoo.org>