maint/bump-generic: more carefully match existing version/template
The old logic was deficient for the following case:
"""
The package exists but no version matching the constraint
(e.g. <x-1.21.x) because it's new in a series and we're bumping
the old one (in this case, gstreamer introduced adaptivedemux in
1.21.x and we're bumping 1.20.x).
"""

Check if the target version has the thing we're going to copy.

Signed-off-by: Sam James <sam@gentoo.org>